refactor: consolidate get_component_from_usage_key implementations
